Fix names of indexes to be consistent with Albion convention
This changes only the "bootstrap" migrations who's job it is
to bring the database up to the point of the last Albion-generated
migration, which is the point at which this new system takes
over.

The Albion migrations would name the indexes explicitly, and the
name it gave them was of the form idx_TABLE_FIELD1_FIELD2.
However, the Postgres default, if you don't specify a name for the
index, is TABLE_FIELD1_FIELD2_idx.

This commit changes all of the "bootstrap" migrations so that they
follow the Albion convention for index names. This is desired,
because it means that all databases, whether old or new, have
identical schemas.

Going forward, we are going to stick to the Postgres convention.

The scripts were fixed using the sed script included in this
commit, by running it like so:

    sed -i -r -f fixup.sed migrations/generic/0000-*.sql

This fixes TI-2688