Add -User tables for Water and Sewer modules
For example, we now have:

WaterPipe, WaterPipeMemo, WaterPipeResult, AND WaterPipeUser

... and same for Sewer

The -User tables have only the 3 basic fields rowid, MetaID, and ID.
But crucially, they have the runtime_schema flag on them.

I've tested this migration with the COU data, and it all works out
of the box.

I followed the recommendations that Johann Rudolph gave us in his
email from 2018-04-13, including the addition of a Link_Code field
to SewerDiversion.

The crucial things that I did not include from Johann's email,
was the 4 new fields that he wanted in the Swift tables, for
drought monitoring. What we want to do instead, is to add user
tables for Swift too, but I'll leave that as future work.