Remove orm_noclient flag from Contour tables in main DB
This flag was causing the Contour tables to be excluded from
the 'main' schema that CrudServer would send to the front-end.

This would cause issues when the user clicked on the map.

In addition to this fix, I have also altered the Typescript code
so that it no longer panics under this condition.

Fixes INF-802